# Reduce allocations during tag serialization