--- a/lib/datadog/statsd/serialization/tag_serializer.rb
+++ b/lib/datadog/statsd/serialization/tag_serializer.rb
@@ -73,6 +73,7 @@ module Datadog
         end
 
         def escape_tag_content(tag)
+          return tag.to_s unless tag.include?('|')
           tag.to_s.delete('|,')
         end