--- a/lib/datadog/statsd/serialization/tag_serializer.rb
+++ b/lib/datadog/statsd/serialization/tag_serializer.rb
@@ -73,8 +73,9 @@ module Datadog
         end
 
         def escape_tag_content(tag)
-          return tag.to_s unless tag.include?('|')
-          tag.to_s.delete('|,')
+          tag = tag.to_s
+          return tag unless tag.include?('|')
+          tag.delete('|,')
         end
 
         def dd_tags(env = ENV)