The `TagSerializer` aims to remove '|,` from tag values. The current implementation simply calls `#delete` on the tag's value.

`delete` is documented as:
> Returns a copy of self with characters specified by selectors removed

Creating this copy uncoditionally increase the amount of object allocations.

This patch is checking the string for occurrences of `|` first and only leans on delete to remove them if at least one exists in the provided value.

The result is a reduction in allocation that can be measured with the benchmark spec.

Before:
```

ruby 3.3.0 (2023-12-25 revision 5124f9ac75) [x86_64-linux]
Warming up --------------------------------------
             no tags     2.896M i/100ms
         global tags     2.908M i/100ms
          tags Array   177.431k i/100ms
           tags Hash   103.517k i/100ms
tags Array + global tags
                       129.709k i/100ms
tags Hash + global tags
                        82.158k i/100ms
Calculating -------------------------------------
             no tags     28.865M (± 5.7%) i/s -    144.809M in   5.034838s
         global tags     29.178M (± 2.9%) i/s -    148.331M in   5.088018s
          tags Array      1.762M (± 3.0%) i/s -      8.872M in   5.039047s
           tags Hash      1.084M (± 3.5%) i/s -      5.486M in   5.068419s
tags Array + global tags
                          1.293M (± 4.1%) i/s -      6.485M in   5.024175s
tags Hash + global tags
                        886.646k (± 4.5%) i/s -      4.437M in   5.014568s

Comparison:
         global tags: 29178293.0 i/s
             no tags: 28865172.6 i/s - same-ish: difference falls within error
          tags Array:  1762195.1 i/s - 16.56x  slower
tags Array + global tags:  1293024.5 i/s - 22.57x  slower
           tags Hash:  1083813.1 i/s - 26.92x  slower
tags Hash + global tags:   886646.1 i/s - 32.91x  slower

      measure IPS
Calculating -------------------------------------
             no tags     0.000  memsize (     0.000  retained)
                         0.000  objects (     0.000  retained)
                         0.000  strings (     0.000  retained)
         global tags     0.000  memsize (     0.000  retained)
                         0.000  objects (     0.000  retained)
                         0.000  strings (     0.000  retained)
          tags Array   240.000  memsize (     0.000  retained)
                         5.000  objects (     0.000  retained)
                         4.000  strings (     0.000  retained)
           tags Hash   400.000  memsize (     0.000  retained)
                         8.000  objects (     0.000  retained)
                         4.000  strings (     0.000  retained)
tags Array + global tags
                       392.000  memsize (   200.000  retained)
                         6.000  objects (     5.000  retained)
                         4.000  strings (     3.000  retained)
tags Hash + global tags
                       552.000  memsize (   200.000  retained)
                         9.000  objects (     5.000  retained)
                         4.000  strings (     3.000  retained)

Comparison:
             no tags:          0 allocated
         global tags:          0 allocated - same
          tags Array:        240 allocated - Infx more
tags Array + global tags:        392 allocated - Infx more
           tags Hash:        400 allocated - Infx more
tags Hash + global tags:        552 allocated - Infx more
      measure memory

Finished in 42.05 seconds (files took 0.09661 seconds to load)
35 examples, 0 failures
```

After:
```
ruby 3.3.0 (2023-12-25 revision 5124f9ac75) [x86_64-linux]
Warming up --------------------------------------
             no tags     2.829M i/100ms
         global tags     2.950M i/100ms
          tags Array   245.805k i/100ms
           tags Hash   130.647k i/100ms
tags Array + global tags
                       167.383k i/100ms
tags Hash + global tags
                        99.813k i/100ms
Calculating -------------------------------------
             no tags     29.521M (± 2.5%) i/s -    149.917M in   5.081495s
         global tags     29.508M (± 3.3%) i/s -    147.479M in   5.003613s
          tags Array      2.368M (± 6.3%) i/s -     12.044M in   5.108053s
           tags Hash      1.184M (± 8.7%) i/s -      5.879M in   5.004052s
tags Array + global tags
                          1.636M (± 5.7%) i/s -      8.202M in   5.032487s
tags Hash + global tags
                          1.037M (± 3.5%) i/s -      5.190M in   5.012192s

Comparison:
             no tags: 29521056.2 i/s
         global tags: 29507653.9 i/s - same-ish: difference falls within error
          tags Array:  2368372.1 i/s - 12.46x  slower
tags Array + global tags:  1635891.9 i/s - 18.05x  slower
           tags Hash:  1184362.2 i/s - 24.93x  slower
tags Hash + global tags:  1036795.0 i/s - 28.47x  slower

      measure IPS
Calculating -------------------------------------
             no tags     0.000  memsize (     0.000  retained)
                         0.000  objects (     0.000  retained)
                         0.000  strings (     0.000  retained)
         global tags     0.000  memsize (     0.000  retained)
                         0.000  objects (     0.000  retained)
                         0.000  strings (     0.000  retained)
          tags Array   120.000  memsize (     0.000  retained)
                         2.000  objects (     0.000  retained)
                         1.000  strings (     0.000  retained)
           tags Hash   280.000  memsize (     0.000  retained)
                         5.000  objects (     0.000  retained)
                         4.000  strings (     0.000  retained)
tags Array + global tags
                       272.000  memsize (    80.000  retained)
                         3.000  objects (     2.000  retained)
                         1.000  strings (     0.000  retained)
tags Hash + global tags
                       432.000  memsize (   240.000  retained)
                         6.000  objects (     5.000  retained)
                         4.000  strings (     3.000  retained)

Comparison:
             no tags:          0 allocated
         global tags:          0 allocated - same
          tags Array:        120 allocated - Infx more
tags Array + global tags:        272 allocated - Infx more
           tags Hash:        280 allocated - Infx more
tags Hash + global tags:        432 allocated - Infx more
      measure memory

Finished in 42.17 seconds (files took 0.09459 seconds to load)
35 examples, 0 failures
```